backend/requirements.txt
Django>=6.0
djangorestframework
django-cors-headers
python-dotenv
xhtml2pdf
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
Pillow>=10.0.0

